+++ webpack.config.js
@@ -1,0 +1,21 @@
+var path = require('path');
+
+module.exports = {
+    entry: './index.js',
+    output: {
+        path: path.resolve(__dirname, 'dist'),
+        filename: 'bundle.js'
+    },
+    module: {
+        rules: [
+            {
+                test: /\.jsx?$/,
+                loader: 'babel-loader',
+                exclude: /node_modules/
+            }
+        ]
+    },
+    devServer: {
+        contentBase: path.resolve(__dirname, "dist")
+    }
+};