+++ CHANGELOG.md
@@ -3,6 +3,21 @@ Following documentation guidelines available at:
 * https://keepachangelog.com/en/0.3.0/
 * http://semver.org/spec/v2.0.0.html
 
+[0.0.1] - 2018-05-30
+### Added
+- Yarn, Webpack, Babel/ES2015, Webpack-dev-server setup
+- yarn: Webpack
+- file: ./webpack.config.js
+- file: ./package.json
+- files: ./upper.js, ./index.js
+- dir: ./dist/
+- file: ./dist/index.html
+- yarn: babel-core, babel-preset-es2015 babel-preset-react
+- file: ./.babelrc
+- yarn: babel-loader
+- yarn: webpack-dev-server
+- yarn: webpack-cli
+
 ## [0.0.0] - 2018-05-30
 ### Added
 - README